+++ lib/src/module/_compose_module.dart
@@ -1,17 +1,12 @@
 part of 'package:dependencies/dependencies.dart';
 
 class _ComposeModule implements Module {
-  @override
-  final String name;
-  final Iterable<Module> _modules;
+  final Iterable<Module> modules;
 
-  _ComposeModule(Iterable<Module> modules)
-      : _modules = modules,
-        name =
-            "ComposeModule[${modules.map((module) => module.name).join(",")}]";
+  _ComposeModule(this.modules);
 
   @override
   void configure(Binder binder) {
-    _modules.forEach(binder.install);
+    modules.forEach(binder.install);
   }
 }