+++ lib/src/module/_empty_module.dart
@@ -2,10 +2,5 @@ part of 'package:dependencies/dependencies.dart';
 
 class _EmptyModule implements Module {
   @override
-  final String name = "EmptyModule";
-
-  _EmptyModule();
-
-  @override
   void configure(Binder binder) {}
 }