+++ lib/src/module/configurer.dart
@@ -1,0 +1,7 @@
+part of 'package:dependencies/dependencies.dart';
+
+/// Represents an object able to configure bindings.
+abstract class Configurer {
+  /// Used to configure all the dependencies.
+  void configure(Binder binder);
+}