+++ lib/src/registry/injector_registry.dart
@@ -2,17 +2,12 @@ part of 'package:dependencies/dependencies.dart';
 
 /// Dependency injection container static factory.
 abstract class InjectorRegistry {
-
   /// Gets the singleton instance.
   static final InjectorRegistry instance = InjectorRegistry();
 
   /// Creates a new [InjectorRegistry] to keep track of different named [Injector].
   /// To use a singleton call [InjectorRegistry.instance].
   factory InjectorRegistry() => _InjectorRegistryContainer();
-
-  /// Enables overriding of [Injector] instances.
-  void enableOverriding();
-
 
   /// Gets an injector from this [InjectorRegistry].
   Injector get({String name});
@@ -28,21 +23,4 @@ abstract class InjectorRegistry {
 
   /// Cleans up the [Injector] instances.
   void clear();
-}
-
-/// Shortcut to inject dependencies using the default injector
-/// [InjectorRegistry.instance] default [Injector].
-T inject<T>({String name, Map<String, dynamic> params, String injectorName}) {
-  return InjectorRegistry.instance
-      .get(name: injectorName)
-      .get(name: name, params: params);
-}
-
-/// Shortcut to inject optional dependencies using the default injector
-/// [InjectorRegistry.instance] default [Injector].
-Optional<T> injectOptional<T>(
-    {String name, String, Map<String, dynamic> params, String injectorName}) {
-  return InjectorRegistry.instance
-      .get(name: injectorName)
-      .getOptional(name: name, params: params);
 }